+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/deviceevents/GBDeviceEventNotificationControl.java
@@ -9,6 +9,7 @@ public class GBDeviceEventNotificationControl extends GBDeviceEvent {
         UNKNOWN,
         DISMISS,
         DISMISS_ALL,
-        OPEN
+        OPEN,
+        MUTE
     }
 }