+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/externalevents/NotificationListener.java
@@ -9,6 +9,8 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
+import android.content.pm.ApplicationInfo;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.os.PowerManager;
 import android.preference.PreferenceManager;
@@ -18,8 +20,6 @@ import android.support.v4.content.LocalBroadcastManager;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.HashSet;
 
 import nodomain.freeyourgadget.gadgetbridge.GBApplication;
 import nodomain.freeyourgadget.gadgetbridge.model.NotificationSpec;
@@ -36,45 +36,60 @@ public class NotificationListener extends NotificationListenerService {
             = "nodomain.freeyourgadget.gadgetbridge.notificationlistener.action.dismiss_all";
     public static final String ACTION_OPEN
             = "nodomain.freeyourgadget.gadgetbridge.notificationlistener.action.open";
+    public static final String ACTION_MUTE
+            = "nodomain.freeyourgadget.gadgetbridge.notificationlistener.action.mute";
 
     private BroadcastReceiver mReceiver = new BroadcastReceiver() {
         @SuppressLint("NewApi")
         @Override
         public void onReceive(Context context, Intent intent) {
             String action = intent.getAction();
-            if (action.equals(ACTION_OPEN)) {
-                StatusBarNotification[] sbns = NotificationListener.this.getActiveNotifications();
-                int handle = intent.getIntExtra("handle", -1);
-                for (StatusBarNotification sbn : sbns) {
-                    if ((int) sbn.getPostTime() == handle) {
-                        try {
-                            PendingIntent pi = sbn.getNotification().contentIntent;
-                            if (pi != null) {
-                                pi.send();
+            switch (action) {
+                case ACTION_MUTE:
+                case ACTION_OPEN: {
+                    StatusBarNotification[] sbns = NotificationListener.this.getActiveNotifications();
+                    int handle = intent.getIntExtra("handle", -1);
+                    for (StatusBarNotification sbn : sbns) {
+                        if ((int) sbn.getPostTime() == handle) {
+                            if (action.equals(ACTION_OPEN)) {
+                                try {
+                                    PendingIntent pi = sbn.getNotification().contentIntent;
+                                    if (pi != null) {
+                                        pi.send();
+                                    }
+                                } catch (PendingIntent.CanceledException e) {
+                                    e.printStackTrace();
+                                }
+                            } else {
+                                // ACTION_MUTE
+                                LOG.info("going to mute " + sbn.getPackageName());
+                                GBApplication.addToBlacklist(sbn.getPackageName());
                             }
-                        } catch (PendingIntent.CanceledException e) {
-                            e.printStackTrace();
                         }
                     }
+                    break;
                 }
-            } else if (action.equals(ACTION_DISMISS)) {
-                StatusBarNotification[] sbns = NotificationListener.this.getActiveNotifications();
-                int handle = intent.getIntExtra("handle", -1);
-                for (StatusBarNotification sbn : sbns) {
-                    if ((int) sbn.getPostTime() == handle) {
-                        if (GBApplication.isRunningLollipopOrLater()) {
-                            String key = sbn.getKey();
-                            NotificationListener.this.cancelNotification(key);
-                        } else {
-                            int id = sbn.getId();
-                            String pkg = sbn.getPackageName();
-                            String tag = sbn.getTag();
-                            NotificationListener.this.cancelNotification(pkg, tag, id);
+                case ACTION_DISMISS: {
+                    StatusBarNotification[] sbns = NotificationListener.this.getActiveNotifications();
+                    int handle = intent.getIntExtra("handle", -1);
+                    for (StatusBarNotification sbn : sbns) {
+                        if ((int) sbn.getPostTime() == handle) {
+                            if (GBApplication.isRunningLollipopOrLater()) {
+                                String key = sbn.getKey();
+                                NotificationListener.this.cancelNotification(key);
+                            } else {
+                                int id = sbn.getId();
+                                String pkg = sbn.getPackageName();
+                                String tag = sbn.getTag();
+                                NotificationListener.this.cancelNotification(pkg, tag, id);
+                            }
                         }
                     }
+                    break;
                 }
-            } else if (action.equals(ACTION_DISMISS_ALL)) {
-                NotificationListener.this.cancelAllNotifications();
+                case ACTION_DISMISS_ALL:
+                    NotificationListener.this.cancelAllNotifications();
+                    break;
             }
 
         }
@@ -87,6 +102,7 @@ public class NotificationListener extends NotificationListenerService {
         filterLocal.addAction(ACTION_OPEN);
         filterLocal.addAction(ACTION_DISMISS);
         filterLocal.addAction(ACTION_DISMISS_ALL);
+        filterLocal.addAction(ACTION_MUTE);
         LocalBroadcastManager.getInstance(this).registerReceiver(mReceiver, filterLocal);
     }
 
@@ -154,13 +170,24 @@ public class NotificationListener extends NotificationListenerService {
             }
         }
 
-        HashSet<String> blacklist = (HashSet<String>) sharedPrefs.getStringSet("package_blacklist", null);
-        if (blacklist != null && blacklist.contains(source)) {
+        if (GBApplication.blacklist != null && GBApplication.blacklist.contains(source)) {
             return;
         }
 
-        // Set application icons for generic notifications
         NotificationSpec notificationSpec = new NotificationSpec();
+
+        // determinate Source App Name ("Label")
+        PackageManager pm = getPackageManager();
+        ApplicationInfo ai = null;
+        try {
+            ai = pm.getApplicationInfo(source, 0);
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+        }
+        if (ai != null) {
+            notificationSpec.sourceName = (String) pm.getApplicationLabel(ai);
+        }
+        
         switch (source) {
             case "org.mariotaku.twidere":
             case "com.twitter.android":